move all the reqwest related code to a different file
