+++ examples/cli-app-with-reqwest/Cargo.toml
@@ -15,3 +15,4 @@ lazy_static = "1.4.0"
 reqwest = "0.11.16"
 async-trait = "0.1.51"
 tokio = { version = "1.27.0", features = ["full"] }
+yaup = "0.2.0"
