+++ examples/cli-app/Cargo.toml
@@ -12,3 +12,4 @@ futures = "0.3"
 serde = { version="1.0",   features = ["derive"] }
 serde_json = "1.0"
 lazy_static = "1.4.0"
+yaup = "0.2.0"
